backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic[email]==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0

SQLAlchemy==2.0.34
asyncpg==0.29.0
alembic==1.13.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
structlog==24.4.0
psycopg2-binary==2.9.9
